+++ src/class46/Code01_BurstBalloons.java
@@ -1,4 +1,4 @@
-package class45;
+package class46;
 
 public class Code01_BurstBalloons {
