+++ src/class47/Code02_RestoreWays.java
@@ -1,4 +1,4 @@
-package class46;
+package class47;
 
 // 整型数组arr长度为n(3 <= n <= 10^4)，最初每个数字是<=200的正数且满足如下条件：
 // 1. 0位置的要求：arr[0]<=arr[1]
